refactor: Seperate the canvas from the map page and make it a component
